moved except/retry logic to ibm_vpc
